+++ apps/misc.h
@@ -42,7 +42,7 @@ void screen_dump(void);
 
 bool settings_parseline(char* line, char** name, char** value);
 bool clean_shutdown(void);
-int default_event_handler_ex(int event, void (*callback)(void *), void *parameter);
-int default_event_handler(int event);
+long default_event_handler_ex(long event, void (*callback)(void *), void *parameter);
+long default_event_handler(long event);
 
 #endif
